+++ controllers/dogu_manager_external_test.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/cloudogu/cesapp-lib/core"
 	cesmocks "github.com/cloudogu/cesapp-lib/registry/mocks"
+	cesremotemocks "github.com/cloudogu/cesapp-lib/remote/mocks"
 	"github.com/cloudogu/k8s-apply-lib/apply"
 	k8sv1 "github.com/cloudogu/k8s-dogu-operator/api/v1"
 	"github.com/cloudogu/k8s-dogu-operator/controllers"
@@ -37,7 +38,7 @@ import (
 
 type doguManagerWithMocks struct {
 	DoguManager           controllers.DoguManager
-	DoguRemoteRegistry    *mocks.DoguRegistry
+	DoguRemoteRegistry    *cesremotemocks.Registry
 	DoguLocalRegistry     *cesmocks.DoguRegistry
 	ImageRegistry         *mocks.ImageRegistry
 	DoguRegistrator       *mocks.DoguRegistrator
@@ -77,6 +78,8 @@ func getDoguManagerWithMocks() doguManagerWithMocks {
 
 	k8sClient := fake.NewClientBuilder().WithScheme(scheme).Build()
 	doguRemoteRegistry := &mocks.DoguRegistry{}
+	resourceGenerator := resource.NewResourceGenerator(scheme)
+	doguRemoteRegistry := &cesremotemocks.Registry{}
 	doguLocalRegistry := &cesmocks.DoguRegistry{}
 	imageRegistry := &mocks.ImageRegistry{}
 	doguRegistrator := &mocks.DoguRegistrator{}
@@ -198,7 +201,7 @@ func TestDoguManager_Install(t *testing.T) {
 		managerWithMocks := getDoguManagerWithMocks()
 		image = &mocks.Image{}
 		image.On("ConfigFile").Return(imageConfig, nil)
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 		managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 		managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 		managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -278,7 +281,7 @@ func TestDoguManager_Install(t *testing.T) {
 		managerWithMocks := getDoguManagerWithMocks()
 		image = &mocks.Image{}
 		image.On("ConfigFile").Return(imageConfig, nil)
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 		managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 		managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 		managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -360,7 +363,7 @@ func TestDoguManager_Install(t *testing.T) {
 	managerRun(t, "failed to validate dependencies", func(t *testing.T) {
 		// given
 		managerWithMocks := getDoguManagerWithMocks()
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 		managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(assert.AnError)
 		_ = managerWithMocks.DoguManager.Client.Create(ctx, ldapCr)
 
@@ -376,7 +379,7 @@ func TestDoguManager_Install(t *testing.T) {
 	managerRun(t, "failed to register dogu", func(t *testing.T) {
 		// given
 		managerWithMocks := getDoguManagerWithMocks()
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 		managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(assert.AnError)
 		managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
 		_ = managerWithMocks.DoguManager.Client.Create(ctx, ldapCr)
@@ -393,7 +396,7 @@ func TestDoguManager_Install(t *testing.T) {
 	managerRun(t, "failed to handle dogu secrets from setup", func(t *testing.T) {
 		// given
 		managerWithMocks := getDoguManagerWithMocks()
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 		managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 		managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
 		managerWithMocks.DoguSecretHandler.On("WriteDoguSecretsToRegistry", mock.Anything, mock.Anything).Return(assert.AnError)
@@ -412,7 +415,7 @@ func TestDoguManager_Install(t *testing.T) {
 	managerRun(t, "failed to create service accounts", func(t *testing.T) {
 		// given
 		managerWithMocks := getDoguManagerWithMocks()
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 		managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 		managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
 		managerWithMocks.DoguSecretHandler.On("WriteDoguSecretsToRegistry", mock.Anything, mock.Anything).Return(nil)
@@ -445,7 +448,7 @@ func TestDoguManager_Install(t *testing.T) {
 	managerRun(t, "error get dogu", func(t *testing.T) {
 		// given
 		managerWithMocks := getDoguManagerWithMocks()
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(nil, assert.AnError)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(nil, assert.AnError)
 
 		_ = managerWithMocks.DoguManager.Client.Create(ctx, ldapCr)
 
@@ -461,7 +464,7 @@ func TestDoguManager_Install(t *testing.T) {
 	managerRun(t, "error on pull image", func(t *testing.T) {
 		// given
 		managerWithMocks := getDoguManagerWithMocks()
-		managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+		managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 		managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(nil, assert.AnError)
 		managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 		managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -484,7 +487,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -513,7 +516,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -543,7 +546,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -574,7 +577,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -605,7 +608,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -637,7 +640,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -669,7 +672,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)
@@ -702,7 +705,7 @@ func TestDoguManager_Install(t *testing.T) {
 			managerWithMocks := getDoguManagerWithMocks()
 			image = &mocks.Image{}
 			image.On("ConfigFile").Return(imageConfig, nil)
-			managerWithMocks.DoguRemoteRegistry.On("GetDogu", mock.Anything).Return(ldapDogu, nil)
+			managerWithMocks.DoguRemoteRegistry.On("Get", mock.Anything).Return(ldapDogu, nil)
 			managerWithMocks.ImageRegistry.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
 			managerWithMocks.DoguRegistrator.On("RegisterDogu", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 			managerWithMocks.DependencyValidator.On("ValidateDependencies", mock.Anything).Return(nil)